--- stock-service/src/main/java/com/destiny/stockservice/application/dto/StockCancelSuccessEvent.java
@@ -1,7 +1,0 @@
-package com.destiny.stockservice.application.dto;
-
-import java.util.UUID;
-
-public record StockCancelSuccessEvent(
-    UUID sagaId
-) { }
