--- stock-service/src/main/java/com/destiny/stockservice/application/dto/StockReservationCancelEvent.java
@@ -1,10 +1,0 @@
-package com.destiny.stockservice.application.dto;
-
-import java.util.List;
-import java.util.UUID;
-
-public record StockReservationCancelEvent(
-    UUID sagaId,
-    UUID orderId,
-    List<StockReservationCancelItem> items
-) { }
