--- stock-service/src/main/java/com/destiny/stockservice/application/dto/StockReservationCancelItem.java
@@ -1,8 +1,0 @@
-package com.destiny.stockservice.application.dto;
-
-import java.util.UUID;
-
-public record StockReservationCancelItem(
-    UUID productId,
-    Integer orderedQuantity
-) { }
