+++ stock-service/src/main/java/com/destiny/stockservice/application/dto/stock/cancel/StockReservationCancelEvent.java
@@ -1,0 +1,15 @@
+package com.destiny.stockservice.application.dto.stock.cancel;
+
+import java.util.List;
+import java.util.UUID;
+
+public record StockReservationCancelEvent(
+    UUID sagaId,
+    UUID orderId,
+    List<StockReservationCancelItem> items
+) {
+    public record StockReservationCancelItem(
+        UUID productId,
+        Integer orderedQuantity
+    ) { }
+}
