+++ stock-service/src/main/java/com/destiny/stockservice/application/dto/stock/reservation/StockReservationFailEvent.java
@@ -1,4 +1,4 @@
-package com.destiny.stockservice.application.dto;
+package com.destiny.stockservice.application.dto.stock.reservation;
 
 import java.util.UUID;
