+++ stock-service/src/main/java/com/destiny/stockservice/application/dto/stock/reservation/StockReservationSuccessEvent.java
@@ -1,4 +1,4 @@
-package com.destiny.stockservice.application.dto;
+package com.destiny.stockservice.application.dto.stock.reservation;
 
 import java.util.List;
 import java.util.UUID;
